+++ app/api/extensions/skatteverket/agi/kvittenser/cron/route.ts
@@ -2,14 +2,12 @@ import { createClient } from '@supabase/supabase-js'
 import { NextResponse } from 'next/server'
 import { ensureInitialized } from '@/lib/init'
 import { verifyCronSecret } from '@/lib/auth/cron'
-import { agiGetKvittenser } from '@/extensions/general/skatteverket/lib/agi-client'
 import { SkatteverketAuthError } from '@/extensions/general/skatteverket/lib/api-client'
 import { markNeedsReconsent, RECONSENT_ERROR_CODES } from '@/extensions/general/skatteverket/lib/token-store'
-import { sendKvittensNotification } from '@/extensions/general/skatteverket/lib/kvittens-notification'
-import { resolveReadAuth, currentSkvEnvironment } from '@/extensions/general/skatteverket/lib/resolve-auth'
+import { currentSkvEnvironment } from '@/extensions/general/skatteverket/lib/resolve-auth'
 import { markGrantRevoked } from '@/extensions/general/skatteverket/lib/connection-store'
-import { formatRedovisare, formatRedovisningsperiod } from '@/lib/skatteverket/format'
-import { completeTaxDeadline } from '@/lib/deadlines/complete-tax-deadline'
+import { reconcileAgiDeclaration } from '@/extensions/general/skatteverket/lib/agi-kvittens-reconcile'
+import { formatRedovisningsperiod } from '@/lib/skatteverket/format'
 import { hasCapability } from '@/lib/entitlements/has-capability'
 import { CAPABILITY } from '@/lib/entitlements/keys'
 
@@ -107,158 +105,25 @@ export async function GET(request: Request) {
     }
 
     try {
-      // Auth resolution prefers system credentials (verified lasombud grant)
-      // and falls back to the company's user token: kvittens polling is the
-      // canonical case for the hybrid model, since the user signed at SKV
-      // and their 65-minute session is usually long dead by the time the
-      // kvittens exists.
-      const resolved = await resolveReadAuth(supabase, companyId, { requires: 'lasombud' })
-      if (!resolved.ok) {
-        if (resolved.reason === 'needs_reconsent') {
-          // A connection flagged needs_reconsent cannot heal on its own
-          // (SKV's per-flow refresh tokens live 65 minutes): skip quietly
-          // instead of failing the same declaration every run.
-          results.push({ declarationId, companyId, period, status: 'expired_token', error: 'needs_reconsent' })
-        } else {
-          results.push({ declarationId, companyId, period, status: 'no_token' })
-        }
-        continue
-      }
-
-      const { data: settings } = await supabase
-        .from('company_settings')
-        .select('org_number, entity_type')
-        .eq('company_id', companyId)
-        .single()
-
-      if (!settings?.org_number) {
-        results.push({ declarationId, companyId, period, status: 'no_company_settings' })
-        continue
-      }
-
-      const arbetsgivare = formatRedovisare(
-        settings.org_number as string,
-        settings.entity_type as 'enskild_firma' | 'aktiebolag',
+      // The shared reconciler resolves auth (system grant → user token),
+      // fetches the kvittens, and on a hit promotes the declaration +
+      // stamps salary_runs / deadline / notification. Auth errors propagate
+      // to the catch below, which owns the cron-specific side effects.
+      const outcome = await reconcileAgiDeclaration(
+        supabase,
+        {
+          id: declarationId,
+          company_id: companyId,
+          salary_run_id: (decl.salary_run_id as string | null) ?? null,
+          period_year: decl.period_year as number,
+          period_month: decl.period_month as number,
+        },
+        { reconciledBy: 'cron' },
       )
 
-      const kvittRes = await agiGetKvittenser(resolved.auth, arbetsgivare, period)
-      if (!kvittRes.ok) {
-        results.push({
-          declarationId, companyId, period,
-          status: 'error',
-          error: kvittRes.error,
-        })
-        continue
-      }
-
-      const kvittens = kvittRes.data.kvittenser?.[0]
-      if (!kvittens?.uuidKvittens) {
-        results.push({ declarationId, companyId, period, status: 'still_pending' })
-        continue
-      }
-
-      // The presence of uuidKvittens confirms SKV signed and accepted
-      // the AGI. signeradTid is the precise signing moment; if SKV omits
-      // it we fall back to reconciliation time + warn so the discrepancy
-      // is investigable. Leaving NULL would hide that the filing occurred
-      // at all, which itself misstates behandlingshistorik (BFNAR 2013:2
-      // kap 8 / BFL 5 kap 6§). The fallback only applies on this code
-      // path because we're inside the kvittens-found branch above.
-      const submittedAt = kvittens.signeradTid || new Date().toISOString()
-      if (!kvittens.signeradTid) {
-        console.warn('[agi-kvittenser-cron] kvittens missing signeradTid; using reconciliation time', {
-          declarationId, companyId, period, uuidKvittens: kvittens.uuidKvittens,
-        })
-      }
-
-      // submitted_by is the token-owning auth.users row: the human who
-      // connected via BankID. The legally load-bearing signer identity
-      // is kvittens.signeradAv (a personnummer), which the token user_id
-      // does NOT necessarily match (e.g. if the connected user is a
-      // bookkeeper but the deklarationsombud signed). We preserve the
-      // full kvittens in response_data so the audit trail (BFL 5 kap 6§,
-      // BFNAR 2013:2 kap 8) records the actual BankID signer regardless
-      // of who triggered the reconciliation.
-      await supabase
-        .from('agi_declarations')
-        .update({
-          status: 'submitted',
-          kvittensnummer: kvittens.uuidKvittens,
-          submitted_at: submittedAt,
-          submitted_by: resolved.tokenUserId,
-          response_data: {
-            signeradAv: kvittens.signeradAv ?? null,
-            signeradTid: kvittens.signeradTid ?? null,
-            uuidKvittens: kvittens.uuidKvittens,
-            arbetsgivare: kvittens.arbetsgivare ?? null,
-            period: kvittens.period ?? null,
-            underlag: kvittens.underlag ?? null,
-            reconciledBy: 'cron',
-          },
-        })
-        .eq('id', declarationId)
-
-      if (decl.salary_run_id) {
-        await supabase
-          .from('salary_runs')
-          .update({ agi_submitted_at: submittedAt })
-          .eq('id', decl.salary_run_id)
-          .eq('company_id', companyId)
-      }
-
-      // Clear the locally-cached submission state so the panel doesn't
-      // pop a stale "awaiting signature" view if the user revisits.
-      await supabase
-        .from('extension_data')
-        .delete()
-        .eq('company_id', companyId)
-        .eq('extension_id', 'skatteverket')
-        .eq('key', `agi_submission_${period}`)
-
-      // The declaration is already flipped to submitted above, and the next
-      // run only revisits pending_signature rows: from here on everything is
-      // best-effort. Each step gets its own try/catch so a failure is logged
-      // as a warning without masking the successful filing or skipping the
-      // remaining confirmation steps.
-
-      // The kvittens is the canonical filing receipt: confirm the period's
-      // arbetsgivardeklaration deadline (terminal state).
-      try {
-        await completeTaxDeadline(
-          supabase,
-          companyId,
-          ['arbetsgivardeklaration'],
-          `${decl.period_year}-${String(decl.period_month).padStart(2, '0')}`,
-          'confirmed'
-        )
-      } catch (deadlineErr) {
-        console.warn('[agi-kvittenser-cron] completeTaxDeadline failed after successful filing', {
-          declarationId, companyId, period,
-          message: deadlineErr instanceof Error ? deadlineErr.message : 'Unknown error',
-        })
-      }
-
-      // Tell the user: signing happened at Skatteverket, often long after
-      // they closed our tab, so this is the only confirmation they get.
-      if (resolved.tokenUserId) {
-        try {
-          await sendKvittensNotification(supabase, {
-            companyId,
-            userId: resolved.tokenUserId,
-            kind: 'agi',
-            period,
-            kvittensnummer: kvittens.uuidKvittens,
-            referenceId: declarationId,
-          })
-        } catch (notifyErr) {
-          console.warn('[agi-kvittenser-cron] sendKvittensNotification failed after successful filing', {
-            declarationId, companyId, period,
-            message: notifyErr instanceof Error ? notifyErr.message : 'Unknown error',
-          })
-        }
-      }
-
-      results.push({ declarationId, companyId, period, status: 'signed' })
+      const result: Result = { declarationId, companyId, period, status: outcome.status }
+      if ('error' in outcome) result.error = outcome.error
+      results.push(result)
     } catch (err) {
       const message = err instanceof Error ? err.message : 'Unknown error'
